+++ pallets/delegation/Cargo.toml
@@ -61,7 +61,7 @@ std = [
   "frame-support/std",
   "frame-system/std",
   "kilt-primitives/std",
-  "kilt-traits/std",
+  "kilt-support/std",
   "log/std",
   "sp-core/std",
   "sp-io/std",
@@ -72,5 +72,5 @@ std = [
 try-runtime = [
   "frame-support/try-runtime",
   "kilt-primitives/try-runtime",
-  "kilt-traits/try-runtime",
+  "kilt-support/try-runtime",
 ]
