+++ pallets/delegation/src/migrations.rs
@@ -17,7 +17,7 @@
 // If you feel like getting in touch with us, you can do so at info@botlabs.org
 
 use codec::{Decode, Encode};
-use kilt_traits::VersionMigratorTrait;
+use kilt_support::traits::VersionMigratorTrait;
 use sp_runtime::traits::Zero;
 
 use crate::*;
