+++ pallets/did/src/deprecated/v2.rs
@@ -17,10 +17,36 @@
 // If you feel like getting in touch with us, you can do so at info@botlabs.org
 
 use codec::{Decode, Encode};
+
+use crate::*;
 use kilt_primitives::Hash;
 
-use crate::{deprecated::Url, *};
 use sp_std::vec::Vec;
+
+type UrlPayload = Vec<u8>;
+
+/// Supported URLs.
+#[derive(Clone, Decode, Encode, PartialEq)]
+pub(crate) enum Url {
+	Http(HttpUrl),
+	Ftp(FtpUrl),
+	Ipfs(IpfsUrl),
+}
+
+#[derive(Clone, Decode, Encode, PartialEq)]
+pub(crate) struct IpfsUrl {
+	pub(crate) payload: UrlPayload,
+}
+
+#[derive(Clone, Decode, Encode, PartialEq)]
+pub(crate) struct FtpUrl {
+	pub(crate) payload: UrlPayload,
+}
+
+#[derive(Clone, Decode, Encode, PartialEq)]
+pub(crate) struct HttpUrl {
+	pub(crate) payload: UrlPayload,
+}
 
 #[derive(Clone, Decode, Encode, PartialEq, Eq)]
 pub(crate) enum ContentType {
