+++ primitives/Cargo.toml
@@ -8,8 +8,6 @@ version = "0.1.0"
 codec = {package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive"]}
 serde = {version = "1.0.101", optional = true, features = ["derive"]}
 
-kilt-traits = {default-features = false, path = "../traits"}
-
 frame-support = {git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.10"}
 sp-core = {git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.10"}
 sp-runtime = {git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.10"}
@@ -22,7 +20,6 @@ fast-gov = []
 std = [
   "codec/std",
   "frame-support/std",
-  "kilt-traits/std",
   "serde",
   "sp-consensus-aura/std",
   "sp-core/std",
@@ -31,5 +28,4 @@ std = [
 ]
 try-runtime = [
   "frame-support/try-runtime",
-  "kilt-traits/try-runtime",
 ]
