+++ primitives/src/lib.rs
@@ -38,7 +38,6 @@ use sp_std::vec::Vec;
 use serde::{Deserialize, Serialize};
 
 pub mod constants;
-pub mod migrations;
 
 /// Opaque types. These are used by the CLI to instantiate machinery that don't
 /// need to know the specifics of the runtime. They can then be made to be
