+++ support/Cargo.toml
@@ -8,12 +8,21 @@ version = "0.26.0"
 
 [dependencies]
 codec = {default-features = false, features = ["derive"], package = "parity-scale-codec", version = "2.0.0"}
+frame-support = {branch = "polkadot-v0.9.10", default-features = false, git = "https://github.com/paritytech/substrate"}
+sp-runtime = {branch = "polkadot-v0.9.10", default-features = false, git = "https://github.com/paritytech/substrate"}
+sp-std = {branch = "polkadot-v0.9.10", default-features = false, git = "https://github.com/paritytech/substrate"}
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
 
 [features]
 default = ["std"]
+try-runtime = [
+  "frame-support/try-runtime"
+]
 std = [
   "codec/std",
+  "frame-support/std",
+  "sp-runtime/std",
+  "sp-std/std"
 ]
