+++ support/src/lib.rs
@@ -18,4 +18,5 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod deposit;
+pub mod migrations;
 pub mod traits;
