--- traits/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-authors = ["KILT <info@kilt.io>"]
-description = "Traits used across the KILT pallets"
-edition = "2018"
-name = "kilt-traits"
-repository = "https://github.com/KILTprotocol/mashnet-node"
-version = "0.26.0"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-frame-support = {git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.10"}
-
-[features]
-default = ["std"]
-std = [
-  "frame-support/std"
-]
-try-runtime = [
-  "frame-support/try-runtime"
-]
